Refactor player type code
