+++ game/player.py
@@ -1,5 +1,5 @@
 from .game_board import GameBoard
-from config import config
+from .player_types import *
 
 import logging
 import os
@@ -46,6 +46,38 @@ class Player:
 
         self.move_dict = config.MOVE_DICT
         self.available_moves = list(self.move_dict.values())
+
+    def play_game(self):
+
+        while True:
+
+            self.logger.info(self.game_board)
+
+            if self._game_is_over():
+                self._save_game_data()
+                self._reset()
+                self.logger.info("Thanks for playing!")
+                break
+
+            move = config.PLAYER_INFO[self.player_type]['func'](self)
+
+            if move == 'quit':
+                self._reset()
+                self.logger.info("Thanks for playing!")
+                break
+
+            if move not in self.move_dict.values():
+                self.logger.warning("Not a recognized move; valid moves are 0,1,2, or 3")
+
+            else:
+                valid_move = self.game_board.move(move)
+
+                self.logger.debug(self.game_board)
+
+                if valid_move:
+                    self._update_game_info()
+                else:
+                    self.available_moves.remove(move)
 
     def _get_max_version(self, default_val = -1):
