--- game/player_types/__init__.py
@@ -1,3 +1,0 @@
-
-from .human import *
-from .bot_v1 import *
