+++ requirements.txt
@@ -1,2 +1,3 @@
 numpy==1.26.4
 prompt_toolkit==3.0.43
+tqdm==4.66.2
